feat: add select statistics
