+++ backend/src/main/java/com/ssafy/attraction/controller/AttractionController.java
@@ -31,4 +31,5 @@ public class AttractionController {
        }
 
     }
+
 }
