+++ backend/src/main/java/com/ssafy/member/model/dao/MemberRepository.java
@@ -11,6 +11,7 @@ public interface MemberRepository {
 
    void register(Member member) throws SQLException;
    Member select(String id) throws SQLException;
+   int selectCountOfMember() throws SQLException;
    void update(Member member) throws SQLException;
    void delete(int memberId) throws SQLException;
