+++ backend/src/main/java/com/ssafy/member/model/service/MemberService.java
@@ -1,18 +1,41 @@
 package com.ssafy.member.model.service;
-
-
-
+import com.ssafy.member.model.dao.MemberRepository;
 import com.ssafy.member.model.dto.Member;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.sql.SQLException;
 
+@Transactional(readOnly = true)
+@Service
+@RequiredArgsConstructor
+public class MemberService{
 
-public interface MemberService {
+    private final MemberRepository memberRepository;
 
-   void register(Member member) throws SQLException;
-   Member login(String id, String password) throws SQLException;
-   void update(Member member) throws SQLException;
-   void delete(int memberId) throws SQLException;
+    public int selectCountOfMember() throws SQLException {
+        return memberRepository.selectCountOfMember();
+    }
+    @Transactional
+    public void register(Member member) throws SQLException {
+        memberRepository.register(member);
+    }
 
+    public Member login(String id, String password) throws SQLException {
+        Member member = memberRepository.select(id);
+        if(member.getPassword().equals(password))
+            return member;
+        return null;
+    }
 
+    @Transactional
+    public void update(Member member) throws SQLException {
+        memberRepository.update(member);
+    }
+
+    @Transactional
+    public void delete(int memberId) throws SQLException {
+        memberRepository.delete(memberId);
+    }
 }
