+++ backend/src/main/resources/mapper/attraction.xml
@@ -5,6 +5,7 @@
 
 <mapper namespace="com.ssafy.attraction.model.dao.AttractionRepository">
 	<resultMap id="attractionResponse" type="Attraction">
+		<result column="content_id" property="contentId"/>
 		<result column="content_type_id" property="contentTypeId"/>
 		<result column="title" property="title"/>
 		<result column="addr1" property="addr1"/>
@@ -22,32 +23,39 @@
 	</resultMap>
 
 	<!--	게시글 목록-->
-<!--	<select id="select" parameterType="SearchCondition" resultMap="attractionResponse">-->
-<!--		SELECT-->
-<!--		    content_type_id,-->
-<!--		    title,-->
-<!--		   	addr1,-->
-<!--		   	addr2,-->
-<!--		   	zipcode,-->
-<!--		   	tel,-->
-<!--		   	first_image,-->
-<!--		  	first_image2,-->
-<!--		   	readCount,-->
-<!--		   	sido_code,-->
-<!--		   	gugun_code,-->
-<!--		   	latitude,-->
-<!--		   	longitude,-->
-<!--		   	mlevel-->
-<!--		FROM attraction_info-->
-<!--		WHERE 1=1-->
-<!--		    <if test="sidoCode != null">-->
-<!--				AND sido_code = #{sidoCode}-->
-<!--			</if>-->
-<!--			<if test="gugunCode != null">-->
-<!--				AND gugun_code = #{gugunCode}-->
-<!--			</if>-->
-<!--			<if test="contentTypeId != null">-->
-<!--				AND content_type_id = #{contentTypeId}-->
-<!--			</if>-->
-<!--	</select>-->
+	<select id="select" parameterType="SearchCondition" resultMap="attractionResponse">
+		SELECT
+		    content_id,
+		    content_type_id,
+		    title,
+		   	addr1,
+		   	addr2,
+		   	zipcode,
+		   	tel,
+		   	first_image,
+		  	first_image2,
+		   	readCount,
+		   	sido_code,
+		   	gugun_code,
+		   	latitude,
+		   	longitude,
+		   	mlevel
+		FROM attraction_info
+		WHERE 1=1
+		    <if test="sidoCode != null">
+				AND sido_code = #{sidoCode}
+			</if>
+			<if test="gugunCode != null">
+				AND gugun_code = #{gugunCode}
+			</if>
+			<if test="contentTypeId != null">
+				AND content_type_id = #{contentTypeId}
+			</if>
+	</select>
+	
+	<select id="selectCountOfAttraction" resultType="int">
+		SELECT COUNT(*) as count
+		FROM attraction_info
+	</select>
+	
 </mapper>
